requirements.txt
streamlit
torch
torchvision
tqdm
pillow
numpy
kagglehub
ultralytics
opencv-python
